+++ packages/dip/src/sibling.ts
@@ -10,12 +10,9 @@ import { ObjectBuilder } from 'typescript-object-builder'
 import { ApiPromise, WsProvider } from '@polkadot/api'
 import { u8aToHex } from '@polkadot/util'
 import type { KeyringPair } from '@polkadot/keyring/types'
-import type { Call, Hash, ReadProof } from '@polkadot/types/interfaces'
-import type { Option } from '@polkadot/types-codec'
-import type { Codec } from '@polkadot/types-codec/types'
+import type { Call, Hash } from '@polkadot/types/interfaces'
 
 import type {
-  Did,
   DidDocument,
   SignatureVerificationRelationship,
   SignerInterface,
@@ -25,213 +22,17 @@ import type {
 import type { PalletDidLookupLinkableAccountLinkableAccountId } from '@kiltprotocol/augment-api'
 
 import { toChain } from '@kiltprotocol/did'
-import { Signers } from '@kiltprotocol/utils'
 
-type ProviderStateRootProofOpts = {
-  providerApi: ApiPromise
-  relayApi: ApiPromise
-  // Optional
-  providerBlockHeight?: number
-}
-type ProviderStateRootProofRes = {
-  proof: ReadProof
-  providerBlockHeight: number
-  relayBlockHeight: number
-}
-async function generateProviderStateRootProof({
-  providerApi,
-  relayApi,
-  // Optional
-  providerBlockHeight,
-}: ProviderStateRootProofOpts): Promise<ProviderStateRootProofRes> {
-  const providerChainId = await providerApi.query.parachainInfo.parachainId()
-  const [providerBlockNumber, providerBlockHash] = await (async () => {
-    if (providerBlockHeight !== undefined) {
-      const blockHash = await providerApi.derive.chain
-        .getBlockByNumber(providerBlockHeight)
-        .then((b) => b.hash)
-      return [providerBlockHeight, blockHash]
-    }
-    const providerLastFinalizedBlockHash =
-      await providerApi.rpc.chain.getFinalizedHead()
-    const providerLastFinalizedBlockHeight = await providerApi.rpc.chain
-      .getHeader(providerLastFinalizedBlockHash)
-      .then((h) => h.number.toNumber())
-    return [providerLastFinalizedBlockHeight, providerLastFinalizedBlockHash]
-  })()
-  const relayParentBlockNumber = await providerApi
-    .at(providerBlockHash)
-    .then((api) => api.query.parachainSystem.lastRelayChainBlockNumber())
-  const relayParentBlockHash = await relayApi.rpc.chain.getBlockHash(
-    relayParentBlockNumber
-  )
+import type { FullnodeAddress } from './utils.js'
+import {
+  defaultValues,
+  generateDipCommitmentProof,
+  generateDipDidSignature,
+  generateDipIdentityProof,
+  generateProviderStateRootProof,
+} from './utils.js'
 
-  const proof = await relayApi.rpc.state.getReadProof(
-    [relayApi.query.paras.heads.key(providerChainId)],
-    relayParentBlockHash
-  )
-
-  return {
-    proof,
-    providerBlockHeight: providerBlockNumber,
-    relayBlockHeight: relayParentBlockNumber.toNumber(),
-  }
-}
-
-type DipCommitmentProofOpts = {
-  did: Did
-  providerApi: ApiPromise
-  providerBlockHash: Hash
-  version: number
-}
-type DipCommitmentProofRes = {
-  proof: ReadProof
-}
-async function generateDipCommitmentProof({
-  did,
-  providerApi,
-  providerBlockHash,
-  version,
-}: DipCommitmentProofOpts): Promise<DipCommitmentProofRes> {
-  const proof = await providerApi.rpc.state.getReadProof(
-    [
-      providerApi.query.dipProvider.identityCommitments.key(
-        toChain(did),
-        version
-      ),
-    ],
-    providerBlockHash
-  )
-
-  return { proof }
-}
-
-type DipIdentityProofOpts = {
-  did: Did
-  keyIds: Array<VerificationMethod['id']>
-  includeWeb3Name: boolean
-  linkedAccounts: readonly PalletDidLookupLinkableAccountLinkableAccountId[]
-  providerApi: ApiPromise
-  version: number
-}
-type DipIdentityProofRes = {
-  proof: {
-    blinded: Codec
-    revealed: Codec
-  }
-  root: Hash
-}
-async function generateDipIdentityProof({
-  did,
-  keyIds,
-  includeWeb3Name,
-  linkedAccounts,
-  providerApi,
-  version,
-}: DipIdentityProofOpts): Promise<DipIdentityProofRes> {
-  const proof = (await providerApi.call.dipProvider.generateProof({
-    identifier: toChain(did),
-    version,
-    keys: keyIds.map((keyId) => keyId.substring(1)),
-    accounts: linkedAccounts,
-    shouldIncludeWeb3Name: includeWeb3Name,
-  })) as Option<Codec>
-
-  const okProof = proof.unwrap() as any
-
-  return { ...okProof }
-}
-
-type DipDidSignatureProviderOpts = {
-  didDocument: DidDocument
-  signers: readonly SignerInterface[]
-  verificationRelationship: SignatureVerificationRelationship
-}
-type DipDidSignatureConsumerOpts = {
-  accountIdRuntimeType: string
-  api: ApiPromise
-  blockNumberRuntimeType: string
-  call: Call
-  identityDetailsRuntimeType: string
-  submitterAddress: KeyringPair['address']
-  // Optional
-  blockHeight?: number
-  genesisHash?: Hash
-}
-type DipDidSignatureOpts = {
-  consumer: DipDidSignatureConsumerOpts
-  provider: DipDidSignatureProviderOpts
-}
-type DipDidSignatureRes = {
-  blockNumber: number
-  signature: Uint8Array
-  type: SignerInterface['algorithm']
-}
-async function generateDipDidSignature({
-  provider: { didDocument, signers, verificationRelationship },
-  consumer: {
-    accountIdRuntimeType,
-    api,
-    blockNumberRuntimeType,
-    call,
-    identityDetailsRuntimeType,
-    submitterAddress,
-    // Optional
-    blockHeight,
-    genesisHash,
-  },
-}: DipDidSignatureOpts): Promise<DipDidSignatureRes> {
-  const blockNumber = await (async () => {
-    if (blockHeight !== undefined) {
-      return blockHeight
-    }
-    const n = await api.query.system.number()
-    return n.toNumber()
-  })()
-  const genesis = await (async () => {
-    if (genesisHash !== undefined) {
-      return genesisHash
-    }
-    return api.query.system.blockHash(0)
-  })()
-  const identityDetails = await (async () => {
-    const maybeIdentityDetails = (await api.query.dipConsumer.identityEntries(
-      toChain(didDocument.id)
-    )) as Option<Codec>
-    try {
-      return maybeIdentityDetails.unwrap()
-    } catch {
-      return api.createType(identityDetailsRuntimeType, null)
-    }
-  })()
-
-  const signaturePayload = api
-    .createType(
-      `(Call, ${identityDetailsRuntimeType}, ${accountIdRuntimeType}, ${blockNumberRuntimeType}, Hash)`,
-      [call, identityDetails, submitterAddress, blockNumber, genesis]
-    )
-    .toU8a()
-  const signer = await Signers.selectSigner(
-    signers,
-    Signers.select.byDid(didDocument, { verificationRelationship })
-  )
-  if (signer === undefined) {
-    throw new Error(
-      `No signer found on DID "${didDocument.id}" for relationship "${verificationRelationship}".`
-    )
-  }
-  const signature = await signer.sign({ data: signaturePayload })
-  return {
-    blockNumber,
-    signature,
-    type: signer.algorithm,
-  }
-}
-
-type WsAddress = `ws${string}`
-type FullnodeAddress = WsAddress
-
-export type Input = {
+export type DipSiblingProofInput = {
   call: Call
   consumerWsOrApi: FullnodeAddress | ApiPromise
   didDocument: DidDocument
@@ -254,20 +55,20 @@ export type Input = {
   linkedAccounts?: readonly PalletDidLookupLinkableAccountLinkableAccountId[]
 }
 
-export const ProofBuilder = ObjectBuilder.new<Input>()
+export const DipSiblingProofBuilder = ObjectBuilder.new<DipSiblingProofInput>()
 
 /**
- * Generate a complete DIP proof for the parameters provided.
+ * Generate a complete DIP proof according to the parameters provided, to be used on a consumer chain of which the provider chain is a sibling.
  *
  * @param params The DIP proof params.
- * @param params.call The [[Call]] on the target chain that requires a DIP origin.
+ * @param params.call The [[Call]] on the consumer chain that requires a DIP origin.
  * @param params.consumerWsOrApi The Websocket address or an [[ApiPromise]] instance for the consumer chain.
- * @param params.didDocument The DID Document of the DIP subject that is performing the cross-chain operation.
+ * @param params.didDocument The [[DidDocument] of the DIP subject that is performing the cross-chain operation.
  * @param params.keyIds The verification method IDs of the DID to be revealed in the cross-chain operation.
  * @param params.proofVersion The version of the DIP proof to generate.
  * @param params.providerWsOrApi The Websocket address or an [[ApiPromise]] instance for the provider chain.
  * @param params.relayWsOrApi The Websocket address or an [[ApiPromise]] instance for the parent relay chain.
- * @param params.signers The list of signers to sign the cross-chain transaction.
+ * @param params.signers The list of [[Signers]] to sign the cross-chain transaction.
  * @param params.submitterAddress The address of the tx submitter on the consumer chain.
  * @param params.verificationRelationship The [[SignatureVerificationRelationship]] required for the DIP operation to be authorized on the consumer chain.
  * @param params.blockHeight [OPTIONAL] The block number on the consumer chain to use for the DID signature. If not provided, the latest best block number is used.
@@ -278,8 +79,10 @@ export const ProofBuilder = ObjectBuilder.new<Input>()
  * @param params.identityDetailsRuntimeType [OPTIONAL] The runtime type definition for the `IdentityDetails` on the consumer chain. If not provided, the `Option<u128>` type, representing a simple nonce, is used.
  * @param params.includeWeb3Name [OPTIONAL] Flag indicating whether the generated DIP proof should include the web3name of the DID subject. If not provided, the web3name is not revealed.
  * @param params.linkedAccounts [OPTIONAL] The list of linked accounts to reveal in the generated DIP proof. If not provided, no account is revealed.
+ *
+ * @returns The [[SubmittableExtrinsic]] containing the signed cross-chain operation, that must be submitted by the account specified as the `submitterAddress` parameter.
  */
-export async function generateDipSiblingProof({
+export async function generateDipProofForSibling({
   call,
   consumerWsOrApi,
   didDocument,
@@ -295,13 +98,12 @@ export async function generateDipSiblingProof({
   genesisHash,
   providerBlockHeight,
   // With defaults
-  accountIdRuntimeType = 'AccountId',
-  blockNumberRuntimeType = 'u64',
-  identityDetailsRuntimeType = 'Option<u128>',
-  includeWeb3Name = false,
-  linkedAccounts = [],
-}: Input): Promise<SubmittableExtrinsic> {
-  // TODO: Move into util function
+  accountIdRuntimeType = defaultValues.accountIdRuntimeType,
+  blockNumberRuntimeType = defaultValues.blockNumberRuntimeType,
+  identityDetailsRuntimeType = defaultValues.identityDetailsRuntimeType,
+  includeWeb3Name = defaultValues.includeWeb3Name,
+  linkedAccounts = defaultValues.linkedAccounts,
+}: DipSiblingProofInput): Promise<SubmittableExtrinsic> {
   const relayApi = await (async () => {
     if (typeof relayWsOrApi === 'string') {
       return ApiPromise.create({ provider: new WsProvider(relayWsOrApi) })
